requirement.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.11
django-allauth==0.44.0
django-cleanup==5.2.0
django-crispy-forms==1.11.2
django-phonenumber-field==5.1.0
idna==2.10
install==1.3.4
oauthlib==3.1.0
phonenumbers==8.12.24
Pillow==8.2.0
py3dns==3.2.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
twilio==6.59.0
urllib3==1.26.5
validate-email==1.3
